+++ .idea/dictionaries/gnagaoka.xml
@@ -1,0 +1,7 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="gnagaoka">
+    <words>
+      <w>claes</w>
+    </words>
+  </dictionary>
+</component>
