+++ ck2mods.json
@@ -1,0 +1,14 @@
+[
+  {
+    "name": "CKII Mod Core",
+    "id": "173534802"
+  },
+  {
+    "name": "CKII Mod Appendix I",
+    "id": "128534375"
+  },
+  {
+    "name": "CKII Mod Appendix II",
+    "id": "173547486"
+  }
+]
